+++ src/App.css
@@ -1,7 +1,14 @@
 .App {
   text-align: center;
 }
-
+.error-alert{
+  width: 300px;
+  margin: 10px auto;
+  padding: 10px;
+  background-color: rgb(214, 61, 0);
+  color: #fff;
+  border-radius: 10px;
+}
 .container {
   padding-right: 15px;
   padding-left: 15px;
